Effects.MonadCallStack

Description

Provides an effect for handling exceptions with callstacks.

Since: 0.1

Class

class Monad m => MonadCallStack m where

Typeclass for CallStack effects. The IO instance uses the machinery from annotated-exception. Note that this means the trycatchetc. function from that package should be used, if you are using the IO instance.

Since: 0.1

Methods

throwWithCallStack :: (Exception e, HasCallStack) => e -> m a

Throws an exception with the CallStack.

Since: 0.1

addCallStack :: HasCallStack => m a -> m a

Adds the CallStack to any thrown exceptions.

Since: 0.1

Utils

displayCallStack :: forall e. Exception e => e -> String

Like displayException, except it has extra logic that attempts to display any found CallStacks in a pretty way.

Since: 0.1

Reexports

data CallStack

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

putStrLnWithCallStack :: HasCallStack => String -> IO ()

as a variant of putStrLn that will get its call-site and print it, along with the string given as argument. We can access the call-stack inside putStrLnWithCallStack with callStack.

>>> :{
putStrLnWithCallStack :: HasCallStack => String -> IO ()
putStrLnWithCallStack msg = do
  putStrLn msg
  putStrLn (prettyCallStack callStack)
:}

Thus, if we call putStrLnWithCallStack we will get a formatted call-stack alongside our string.

>>> putStrLnWithCallStack "hello"
hello
CallStack (from HasCallStack):
  putStrLnWithCallStack, called at <interactive>:... in interactive:Ghci...

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.8.1.0

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

data AnnotatedException exception

The AnnotatedException type wraps an exception with a [Annotation]. This can provide a sort of a manual stack trace with programmer provided data.

Since: annotated-exception-0.1.0.0

Constructors

AnnotatedException

Fields

annotations :: [Annotation]

exception :: exception

data Annotation where

An Annotation is a wrapper around a value that includes a Typeable constraint so we can later unpack it. It is essentially a Dynamic, but we also include Show so that you can always fall back to simply showing the Annotation if it is otherwise unrecognized.

Since: annotated-exception-0.1.0.0

Constructors

Annotation :: forall a. AnnC a => a -> Annotation

throw :: forall e m a. (MonadIO m, Exception e, HasCallStack) => e -> m a

Like throw, but uses MonadIO instead of MonadThrow.

Since: annotated-exception-0.1.2.0

try :: forall e m a. (MonadUnliftIO m, Exception e) => m a -> m (Either e a)

Like try but uses MonadUnliftIO instead of MonadCatch.

Since: annotated-exception-0.1.2.0

catch :: forall e m a. (MonadUnliftIO m, Exception e, HasCallStack) => m a -> (e -> m a) -> m a

Like catch, but uses MonadUnliftIO instead of MonadCatch.

Since: annotated-exception-0.1.2.0